set up unit testing file
